// WinWallpaper/NodeJs/wallpaper/vue-wallpaper/src/app/views/index/type.ts
export type UserParams = {
  schemecolor?: AnyValue;
  schemecolor_color?: NumberValue;
  text_img?: AnyValue;
  imgIsRefresh?: BoolValue;
  imgLastActive?: NumberValue;
  imgRefreshTime?: NumberValue;
  text_line?: AnyValue;
  lineStyle?: NumberValue;
  lineMulticolor?: BoolValue;
  lineColorFollow?: BoolValue;
  lineColor?: StringValue;
  text_line1?: AnyValue;
  line1_r?: NumberValue;
  line1_rotation?: NumberValue;
  line1_cX?: NumberValue;
  line1_cY?: NumberValue;
  line1_range?: NumberValue;
  text_line2?: AnyValue;
  line2_range?: NumberValue;
  line2_width?: NumberValue;
  line2_startY?: NumberValue;
  text_modules?: AnyValue;
  modulesMulticolor?: BoolValue;
  modules_time?: AnyValue;
  modules_timeShow?: BoolValue;
  modules_timeFontSize?: NumberValue;
  modules_timeX?: NumberValue;
  modules_timeY?: NumberValue;
  modules_date?: AnyValue;
  modules_dateShow?: BoolValue;
  modules_dateFontSize?: NumberValue;
  modules_dateX?: NumberValue;
  modules_dateY?: NumberValue;
  modules_lunar?: NumberValue;
  modules_lunarShow?: BoolValue;
  modules_lunarFontSize?: NumberValue;
  modules_lunarX?: NumberValue;
  modules_lunarY?: NumberValue;
  modules_gzDate?: NumberValue;
  modules_gzDateShow?: BoolValue;
  modules_gzDateFontSize?: NumberValue;
  modules_gzDateX?: NumberValue;
  modules_gzDateY?: NumberValue;
  modules_weather?: AnyValue;
  modules_weatherShow?: BoolValue;
  modules_weatherFontSize?: NumberValue;
  modules_weatherX?: NumberValue;
  modules_weatherY?: NumberValue;
  modules_city?: StringValue;
  modules_weatherRefresh?: NumberValue;
  monitro_text?: AnyValue;
  monitro_show?: BoolValue;
  monitroMulticolor?: BoolValue;
  monitro_fontSize?: NumberValue;
  monitro_x?: NumberValue;
  monitro_y?: NumberValue;
  doublePY_text?: AnyValue;
  doublePY_show?: BoolValue;
  doublePY_type?: NumberValue;
  doublePY_x?: NumberValue;
  doublePY_y?: NumberValue;
  doublePY_leftStyle?: StringValue;
  doublePY_rightStyle?: StringValue;
  musicPlayer?: AnyValue;
  musicPlayer_show?: BoolValue;
  musicPlayer_Root?: StringValue;
  musicPlayer_x?: NumberValue;
  musicPlayer_y?: NumberValue;
  netWorkSpeekMonitor?: AnyValue;
  netWorkSpeekMonitor_show?: BoolValue;
  netWorkSpeekMonitor_multicolor?: BoolValue;
  netWorkSpeekMonitor_x?: NumberValue;
  netWorkSpeekMonitor_y?: NumberValue;
  netWorkSpeekMonitor_fontSize?: NumberValue;
  NumberValue?: StringValue;
  NumberValue_show?: StringValue;
  danmaku?: AnyValue;
  danmaku_show?: BoolValue;
};

type NumberValue = {
  value: number;
};

type StringValue = {
  value: string;
};
type BoolValue = {
  value: boolean;
};

type AnyValue = {};